+++ backend/src/main/scala/petclinic/services/OwnerService.scala
@@ -1,10 +1,7 @@
 package petclinic.services
 
-import zio.{Task, URLayer, ZEnvironment, ZIO, ZLayer}
-import petclinic.QuillContext
+import zio._
 import petclinic.models._
-
-import javax.sql.DataSource
 
 trait OwnerService {
 
@@ -57,56 +54,4 @@ object OwnerService {
   ): ZIO[OwnerService, Throwable, Unit] =
     ZIO.serviceWithZIO[OwnerService](_.update(id, firstName, lastName, address, phone, email))
 
-}
-
-final case class OwnerServiceLive(dataSource: DataSource) extends OwnerService {
-
-  import QuillContext._
-
-  override def create(firstName: String, lastName: String, address: String, phone: String, email: String): Task[Owner] =
-    for {
-      owner <- Owner.make(firstName, lastName, address, phone, email)
-      _     <- run(query[Owner].insertValue(lift(owner))).provideEnvironment(ZEnvironment(dataSource))
-    } yield owner
-
-  override def delete(id: OwnerId): Task[Unit] =
-    run(query[Owner].filter(_.id == lift(id)).delete)
-      .provideEnvironment(ZEnvironment(dataSource))
-      .unit
-
-  override def get(id: OwnerId): Task[Option[Owner]] =
-    run(query[Owner].filter(_.id == lift(id)))
-      .provideEnvironment(ZEnvironment(dataSource))
-      .map(_.headOption)
-
-  override def getAll: Task[List[Owner]] =
-    run(query[Owner])
-      .provideEnvironment(ZEnvironment(dataSource))
-
-  override def update(
-      id: OwnerId,
-      firstName: Option[String],
-      lastName: Option[String],
-      address: Option[String],
-      phone: Option[String],
-      email: Option[String]
-  ): Task[Unit] =
-    run(
-      dynamicQuery[Owner]
-        .filter(_.id == lift(id))
-        .update(
-          setOpt(_.firstName, firstName),
-          setOpt(_.lastName, lastName),
-          setOpt(_.address, address),
-          setOpt(_.phone, phone),
-          setOpt(_.email, email)
-        )
-    )
-      .provideEnvironment(ZEnvironment(dataSource))
-      .unit
-
-}
-
-object OwnerServiceLive {
-  val layer: URLayer[DataSource, OwnerService] = ZLayer.fromFunction(OwnerServiceLive.apply _)
 }
