+++ backend/src/main/scala/petclinic/services/VetService.scala
@@ -1,9 +1,6 @@
 package petclinic.services
 
-import zio.{Task, URLayer, ZEnvironment, ZIO, ZLayer}
-
-import javax.sql.DataSource
-import petclinic.QuillContext
+import zio._
 import petclinic.models._
 
 trait VetService {
@@ -19,23 +16,4 @@ object VetService {
 
   def get(vetId: VetId): ZIO[VetService, Throwable, Option[Vet]] =
     ZIO.serviceWithZIO[VetService](_.get(vetId))
-}
-
-final case class VetServiceLive(dataSource: DataSource) extends VetService {
-
-  import QuillContext._
-
-  override def getAll: Task[List[Vet]] =
-    run(query[Vet])
-      .provideEnvironment(ZEnvironment(dataSource))
-
-  override def get(vetId: VetId): Task[Option[Vet]] =
-    run(query[Vet].filter(_.id == lift(vetId)))
-      .provideEnvironment(ZEnvironment(dataSource))
-      .map(_.headOption)
-}
-
-object VetServiceLive {
-  val layer: URLayer[DataSource, VetService] =
-    ZLayer.fromFunction(VetServiceLive.apply _)
 }
